requirements.txt
mysql-connector-python==8.0.29
tk

# pip install -r requirements.txt 
#used to install environment for run it.